+++ src/test/java/dev/zayt/atrilha/AtrilhaApplicationTests.java
@@ -2,8 +2,10 @@ package dev.zayt.atrilha;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
 
 @SpringBootTest
+@ActiveProfiles("test")
 class AtrilhaApplicationTests {
     @Test
     void contextLoads() { }